pyproject.toml
[project]
name = "django-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-cotton>=2.6.2",
    "django-stubs[compatible-mypy]>=6.0.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs[compatible-mypy]>=3.16.9",
    "poethepoet>=0.44.0",
]

[tool.poe.tasks]
dev = "python manage.py runserver"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
superuser = "python manage.py createsuperuser"
lint = "ruff check ."
format = "ruff format ."

# * Settings from https://github.com/wemake-services/wemake-django-template

[tool.uv]
# Docs: https://docs.astral.sh/uv/reference/settings
package = false

[tool.ruff]
# Ruff config: https://docs.astral.sh/ruff/settings
preview = true
fix = true
target-version = "py314"
line-length = 120
extend-exclude = ["**/migrations/*.py"]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
# select = [
#   "A",    # flake8-builtins
#   "ASYNC",
#   "B",    # flake8-bugbear
#   "C4",   # flake8-comprehensions
#   "C90",  # maccabe
#   "COM",  # flake8-commas
#   "D",    # pydocstyle
#   "DTZ",  # flake8-datetimez
#   "E",    # pycodestyle
#   "ERA",  # flake8-eradicate
#   "EXE",  # flake8-executable
#   "F",    # pyflakes
#   "FA",   # flake8-future-annotations
#   "FBT",  # flake8-boolean-trap
#   "FLY",  # pyflint
#   "FURB", # refurb
#   "G",    # flake8-logging-format
#   "I",    # isort
#   "ICN",  # flake8-import-conventions
#   "ISC",  # flake8-implicit-str-concat
#   "LOG",  # flake8-logging
#   "N",    # pep8-naming
#   "PERF", # perflint
#   "PIE",  # flake8-pie
#   "PL",   # pylint
#   "PT",   # flake8-pytest-style
#   "PTH",  # flake8-use-pathlib
#   "Q",    # flake8-quotes
#   "RET",  # flake8-return
#   "RSE",  # flake8-raise
#   "RUF",  # ruff
#   "S",    # flake8-bandit
#   "SIM",  # flake8-simpify
#   "SLF",  # flake8-self
#   "SLOT", # flake8-slots
#   "T100", # flake8-debugger
#   "TRY",  # tryceratops
#   "UP",   # pyupgrade
#   "W",    # pycodestyle
#   "YTT",  # flake8-2020
# ]
ignore = [
  "A005",   # allow to shadow stdlib and builtin module names
  # Different doc rules that we don't really care about:
  "D100",
  "D104",
  "D106",
  "D203",
  "D212",
  "D401",
  "D404",
  "D405",
  "ISC001",  # implicit string concat conflicts with `ruff format`
  "ISC003",  # prefer explicit string concat over implicit concat
  "PLR09",   # we have our own complexity rules
  "PLR2004", # do not report magic numbers
  "PLR6301", # do not require classmethod / staticmethod when self not used
  "RUF067",  # wps handles `__init__.py` logic better
  "TRY003",  # long exception messages from `tryceratops`
#   'COM812'
  'RUF012'   # permission_classes = [IsAuthenticated]
]
external = ["WPS"]

[tool.ruff.lint.per-file-ignores]
"/migrations/*.py" = ["D101", "E501", "RUF012"]
"/tests/*.py" = ["S101"]

[tool.mypy]
# Mypy configuration:
# https://mypy.readthedocs.io/en/latest/config_file.html
enable_error_code = [
  "truthy-bool",
  "truthy-iterable",
  "redundant-expr",
  "unused-awaitable",
  "ignore-without-code",
  "possibly-undefined",
  "redundant-self",
  "explicit-override",
  "unimported-reveal",
  "deprecated",
  "exhaustive-match",
]

explicit_package_bases = true
ignore_missing_imports = true
local_partial_types = true
strict = true
strict_bytes = true
warn_unreachable = true

plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = ["*.migrations", "*.migrations.*"]
# Django migrations should not produce any errors (they are tested anyway):
ignore_errors = true

[tool.django-stubs]
django_settings_module = "core.settings"
strict_settings = false
